+++ project/src/main/kotlin/dev/enderman/minecraft/plugins/fire/better/events/listeners/EntityContactListener.kt
@@ -1,20 +1,22 @@
 package dev.enderman.minecraft.plugins.fire.better.events.listeners
 
-import org.bukkit.damage.DamageType
 import org.bukkit.event.EventHandler
 import org.bukkit.event.Listener
 import org.bukkit.event.entity.EntityDamageByEntityEvent
+import org.bukkit.event.entity.EntityDamageEvent.DamageCause
 
-val contactAttacks = listOfNotNull(DamageType.PLAYER_ATTACK, DamageType.MOB_ATTACK, DamageType.MOB_ATTACK_NO_AGGRO)
+val contactAttacks = listOfNotNull(
+    DamageCause.ENTITY_ATTACK,
+    DamageCause.ENTITY_SWEEP_ATTACK,
+    DamageCause.CRAMMING,
+)
 
 class EntityContactListener : Listener {
     @EventHandler fun onHit(event: EntityDamageByEntityEvent) {
         val entity = event.entity
         val damager = event.damager
 
-        val isDirect = event.damageSource.damageType in contactAttacks
-
-        if (!isDirect) return
+        if (event.cause !in contactAttacks) return
 
         val entityOnFire = entity.fireTicks > 0
         val damagerOnFire = damager.fireTicks > 0