+++ project/src/main/kotlin/dev/enderman/minecraft/plugins/fire/better/events/suffocation/IsSuffocationDamage.kt
@@ -1,0 +1,7 @@
+package dev.enderman.minecraft.plugins.fire.better.events.suffocation
+
+import org.bukkit.event.entity.EntityDamageEvent
+
+fun EntityDamageEvent.isSuffocationDamage(): Boolean {
+    return cause == EntityDamageEvent.DamageCause.SUFFOCATION
+}