+++ docs/CHANGELOG.md
@@ -16,6 +16,13 @@
 
 ## Log
 
+### 0.2.0
+
+#### Changes
+
+- Fire protection now protects even better against fire damage. Full fire protection makes you completely immune to fire.
+- Fire protection will sometimes prevent durability loss from fire, with full fire protection preventing all durability loss.
+
 ### 0.1.0
 
 #### Additions
@@ -24,7 +31,10 @@
 - Hitting a mob that is on fire will spread the fire to you, and vice versa.
 - Right-clicking entities with a flint and steel now sets them on fire.
 
+#### Changes
+
+- Attempting to put out a fire with your hand will instead set you on fire.
+
 #### Removals
 
-- Fire can no longer be put out by punching it.
 - Entities on fire do not extinguish naturally.